+++ registration/pubspec.yaml
@@ -12,8 +12,7 @@ dependencies:
   http: ^1.2.2
   crypto: ^3.0.3
   provider: ^6.1.1
-
-  provider: any
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
